Fixed logic error
The fallback assignment converted the value to the type of the current
field value. This type, however, may be a UnionAll or a subtype, and not
the actual data type.
The algorithm has been adjusted to efficiently determine the *correct*
data type.

TODO: Update the entire algorithm to use a similar approach for
virtual properties.